+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   through the introduction of the `remappings` argument to the `launch`
   method.
 * Fixed mishandling of complex parameters when reading rosparam tags.
+* Added `to_xml_string` and `to_xml_file` to `LaunchConfig`.
 
 
 # 1.3.0 (2020-06-02)